[MAR-2794] Reuse one CI candidate and isolate historical checks
