+++ README.md
@@ -221,8 +221,12 @@ Without `root`, Encephalon walks upward to the nearest valid Git repository mark
 node ./scripts/check-generated-version.ts
 bun install --frozen-lockfile --ignore-scripts
 bun run typecheck
+bun run lint
 bun run test
+bun run test:tooling
 bun run build
+bun run test:package
+bun run test:history
 node ./scripts/check-package.ts --retain-tarball package-artifacts
 node ./scripts/check-package-metadata.ts
 node ./scripts/check-release-compatibility.ts package-artifacts/encephalon-0.3.0.tgz
@@ -244,13 +248,15 @@ The full profile runs every operation in fresh child processes with two discarde
 
 `bun run check:generated` is contributor convenience for trusted local checkouts. Its target generated-version checker is non-mutating, but package lifecycle hooks and Bun preloads mean the alias is not an authority for untrusted changes. Every source-building CI job runs the exact direct command `node ./scripts/check-generated-version.ts` immediately after Node setup and before Bun setup or installation so those hooks and preloads cannot repair stale or missing committed source before validation. `check:package` rejects packed paths outside reviewed tracked package inputs and the expected generated distribution, installs the tarball with lifecycle scripts disabled, imports the public API, and runs the bundled CLI using Node. `check:publish` exercises npm's publish-time manifest normalisation without uploading anything. The `bun run check:package` and `bun run check:publish` contributor commands launch those Node scripts, while CI invokes those scripts directly so package lifecycle hooks and Bun preloads cannot interpose.
 
-CI runs four verification lanes: Node 24.15.0 on Ubuntu, macOS, and Windows, plus Node 26 on Ubuntu. After all four succeed, one package job gates the clean source tree around the test suite and explicit build, requires the artifact directory to be absent, then checks and retains `package-artifacts/encephalon-0.3.0.tgz` together with its deterministic metadata sidecar. The sidecar records the exact byte size, SHA-1, SHA-256, SHA-512, npm integrity, package version, source commit, and fixed repository-relative tarball path. CI recomputes that authority before uploading both files as `encephalon-npm-package` on pull requests and trusted pushes to `main`.
+Ordinary `bun run test` runs the complete build-independent source suite. `test:tooling` runs cheap tool checks; `test:package` requires an explicit build. `test:history` runs the historical release-tool process regressions, and `test:all` retains all four groups for full local verification after a build. New test files require an explicit coverage classification. macOS and Windows run filesystem, path and process tests, plus changed source-only tests; package consumers exercise installed executable behaviour on all three operating systems.
 
-Node 24.15.0 and Node 26 candidate jobs build the reviewed checkout, download those same artifact bytes, verify the metadata sidecar, and invoke `check-package.ts --tarball` against the fixed repository-relative tarball path without extracting or repacking it. The final release-equivalent package gate repeats those checks, runs the compatibility oracle once with that exact path, and passes the same path to the tarball-only publish dry run. The published npm oracle requires network access, but no job receives repository, provider, or npm credentials and no job publishes.
+CI checks source on Linux with Node 24.15.0 and Node 26. After both source lanes, lint, types and cheap tool tests succeed, one package job checks the clean tree before and after the explicit build, requires `package-artifacts` to be absent, and retains one tarball and metadata sidecar. The filename comes from the validated package version. The sidecar records size, SHA-1, SHA-256, SHA-512, npm integrity, source commit and repository-relative path. Consumers download the producer's immutable artifact ID, check its separately recorded SHA-256, and restore only the reviewed `dist` files from that tarball. They never rebuild or repack the candidate.
 
-Pull requests therefore upload and retain the auditable candidate instead of keeping it only on the runner. Before a manual, explicit-approval, tarball-only publish from trusted `main`, maintainers must compare the trusted-main tarball and metadata sidecar with the reviewed pull-request candidate and require the tarballs to be byte-identical. Pull-request runs cancel superseded work, while each trusted push has an independent run so a newer push cannot cancel its artifact handoff.
+Actual published `encephalon@0.3.0` upgrade/downgrade compatibility runs for every main candidate and for pull requests touching public, persisted, package or release behaviour. Historical release-tool regressions run when their tools or CI change. The `release` label, a manual main run and the weekly main schedule force both checks. Deletions and both sides of renames participate in selection; unknown paths conservatively require compatibility. Known ordinary documentation or test-only pull requests may skip historical checks. The unchanged six Linux benchmark shards still collect 20 measured samples after two warmups for each operation and side. The exact candidate also receives the benchmark smoke and tarball-only npm publish dry run. No job publishes to npm.
 
-After rollout, branch protection must require exactly `verify (ubuntu-latest)`, `verify (macos-latest)`, `verify (windows-latest)`, `verify (ubuntu-current)`, `Build exact package candidate`, `candidate (Node 24.15.0)`, `candidate (Node 26)`, and `Release-equivalent package gate`. Maintainers can repeat the read-only verification with:
+Pull requests retain the auditable candidate. A separate workflow copies a successful main run's verified artifact only after that CI run has completed successfully; its trusted artifact name includes the source commit, originating run and attempt. Superseded PR and main runs are cancelled, and cancelled runs cannot promote a candidate. Consumer-only reruns reuse the original producer artifact; full reruns create an attempt-specific candidate. Before a manual, explicit-approval publish handoff, require the trusted artifact to match current main and compare its bytes with the reviewed PR tarball.
+
+The required branch-protection contexts remain `verify (ubuntu-latest)`, `verify (macos-latest)`, `verify (windows-latest)`, `verify (ubuntu-current)`, `Release-equivalent package gate` and `Greptile Review`. Greptile must report 5/5 for the reviewed head. Maintainers can repeat the read-only verification with:
 
 ```bash
 gh api repos/isaachinman/encephalon/branches/main/protection/required_status_checks \
