+++ package.json
@@ -55,7 +55,10 @@
     "check:publish": "node ./scripts/check-publish.ts",
     "lint": "ultracite check",
     "prepack": "bun run build && bun run check:package",
-    "test": "node scripts/test-ci.ts runtime",
+    "test": "node scripts/test-ci.ts source",
+    "test:all": "node scripts/test-ci.ts all",
+    "test:history": "node scripts/test-ci.ts history",
+    "test:package": "node scripts/test-ci.ts package",
     "test:tooling": "node scripts/test-ci.ts tooling",
     "typecheck": "tsc --project tsconfig.src.json --noEmit && tsc --project tsconfig.scripts.json --noEmit && tsc --project tsconfig.test.json --noEmit && tsc --project tsconfig.runtime-guard.json --noEmit"
   },
